ch(refactor project) refactor project structure create config.js which contains connection pooling information for dev and test environments modify db.js by using the configuration properties from config.js rename api.v1.js to route.js
[finishes #167276572]
